Create VerticalScrollContainer and split it out of scrollable image gallery
